8월 5일 - 인원 수  +/-
