+++ .idea/misc.xml
@@ -29,6 +29,8 @@
         <entry key="../../../../layout/compose-model-1659557280896.xml" value="0.21875" />
         <entry key="../../../../layout/compose-model-1659647170740.xml" value="0.1" />
         <entry key="../../../../layout/compose-model-1659647226731.xml" value="1.0" />
+        <entry key="../../../../layout/compose-model-1659810339480.xml" value="0.1" />
+        <entry key="../../../../layout/compose-model-1659907240056.xml" value="0.15555555555555556" />
         <entry key="..\:/Users/Min Jae/AndroidStudioProjects/KBSC_Cooperate/app/src/main/res/drawable/ic_person.xml" value="0.1" />
         <entry key="..\:/Users/Min Jae/AndroidStudioProjects/KBSC_Cooperate/app/src/main/res/drawable/ic_plane.xml" value="0.1" />
         <entry key="..\:/Users/Min Jae/AndroidStudioProjects/KBSC_Cooperate/app/src/main/res/font/raleway_light.xml" value="0.1" />
